# frontend/openapi/asbl-club-api.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "asbl.club API",
    "description" : "Public read API exposing associations and their public events. Reads are open with CORS for showcase sites. Everything else requires a JWT access token from POST /api/v1/auth/login, sent as a Bearer token.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/events/{eventId}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get a public event with its association, location and tickets",
        "operationId" : "getEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicEvent"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{eventId}/availability" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Remaining seats per ticket category of a public event",
        "operationId" : "getEventAvailability",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SeatAvailability"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asbls/{slug}" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "Get an association by slug",
        "operationId" : "getAsbl",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AsblResource"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asbls/{slug}/events" : {
      "get" : {
        "tags" : [ "Public" ],
        "summary" : "List an association's public events",
        "operationId" : "listAsblEvents",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EventFeedItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AsblResource" : {
        "required" : [ "denomination", "slug" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          }
        }
      },
...
      "PublicEvent" : {
        "required" : [ "asbl", "id", "startsAt", "tickets", "title" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "location" : {
            "type" : "string"
          },
          "asbl" : {
            "$ref" : "#/components/schemas/AsblResource"
          },
          "cancellableUntil" : {
            "type" : "string",
            "description" : "Until when buyers may cancel their ticket and be refunded; absent: not refundable on request",
            "format" : "date-time"
          },
          "tickets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PublicTicket"
            }
          }
        }
      },
      "PublicTicket" : {
        "required" : [ "id", "label", "price", "remaining" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "label" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number",
            "description" : "Price in euros"
          },
          "remaining" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "SeatAvailability" : {
        "required" : [ "categoryId", "remaining" ],
        "type" : "object",
        "properties" : {
          "categoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remaining" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "EventFeedItem" : {
        "required" : [ "id", "startsAt", "title" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startsAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearer" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}